% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{Ridge regression}
\description{
Returns the result of the Ridge Regression
}
\section{Fields}{

\describe{
\item{\code{formula}}{a formula, format y ~ x_1 + x_2 + ... + x_n.}

\item{\code{data}}{a data frame.}

\item{\code{lambda}}{a hyperparameter.}
}}


\examples{
data(iris)
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda = 0)$print()
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda = 0)$predict()
ridgereg$new(Petal.Length~Sepal.Width+Sepal.Length, data=iris, lambda = 0)$coef()
}
